requirements.txt
web3==6.11.3
python-dotenv==1.0.0
eth-account==0.10.0
py-solc-x==2.0.2
python-telegram-bot==20.7
